Move rest handler of subresources of pod to rest package
